LE-28: WIP: Saving layouts
So far, we have a working means of saving an Oasis layout in JSON
format.

We have introduced an Exporter system, and the usage is quite simple.

All exporters have an interface method

```
public void Export(SettingsData projectSettings, Dictionary<string,
object> layout, string fileName)
```

The export is supplied with the project settings and the layoutb as a
dictionary.

At an abstract level, validation is performed on projectSettings and
layout to ensure that they are correctly formed. This is performed for
all exporters irrespective of target format.

Then the specific exporter will serialize the project settings and
layout and produce an output file.

For now, I have implemented two Exporters, one a MAMEExporter and the
other an OasisExporter

The MAME exporter only calls the base validations for now and does not
have an implementation.

The OASIS exporter outputs a JSON file to a path provided.

I have wired this into the menus as a test for now.