BUGFIX-lamp-name-should-be-unique
Each component, such as a lamp should have a unique identifier, a recent
code change removed the unique identifier from the lamp name

This PR puts the missing component back.

The current implementation of serialization de-deplicates components as
that each one is distinct. Removing the ID from the lamp identifier
causes problems with this deduplication.