LE-MISC-003: Parse images from layout
When parsing a layout for it's representation, we need to parse the
images from the layout also.

This is the first step in that work. We will extract the images from the
existing layout, flip them in the correct way and then store them with a
filename to disk.

TODO: We will need to have a better way of saving layouts so that they
can be stored into a directory by the user.

TODO: I'm not sure why the internal representation of the layout has two
copies of the image (one inverted) but it seems that if we process the
first instance of these images, they can be processed in a predictible
way.